Setting environmental variables
