+++ .github/workflows/test-inputs.yml
@@ -8,7 +8,6 @@ on:
 
 env:
   RUST_BACKTRACE: 1
-  CARGO_TERM_COLOR: always
 
 jobs:
   test-package-patch:
